rm and reinstall dependencies, move tests to src
