+++ package.json
@@ -2,7 +2,6 @@
   "name": "ace",
   "version": "1.0.0",
   "description": "Threejs Typescript tennis game",
-  "main": "index.js",
   "type": "module",
   "scripts": {
     "test": "jest",
@@ -26,7 +25,8 @@
     "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-n": "^15.2.4",
     "eslint-plugin-promise": "^6.0.0",
-    "jest": "^27.0.3",
+    "jest": "^29.0.3",
+    "ts-jest": "^29.0.3",
     "three": "^0.146.0",
     "ts-loader": "^9.4.1",
     "webpack": "^5.75.0",
@@ -39,7 +39,6 @@
     "socket.io": "^4.5.3",
     "socket.io-client": "^4.5.3",
     "three.meshline": "^1.4.0",
-    "ts-jest": "^27.0.3",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.3"
   }
