+++ src/server/Dockerfile
@@ -6,8 +6,6 @@ COPY tsconfig*.json ./
 COPY ./src/server ./app
 
 RUN npm install -g ts-node
-
-# Install dependencies
 RUN yarn install
 
 EXPOSE 3000
