+++ src/test/babel.config.js
