+++ src/test/server/Game.test.ts
@@ -1,4 +1,4 @@
-import { Game } from '../../src/server/Game.ts';
+import { Game } from '../../server/game/Game.ts';
 
 describe('Game increment from', () => {
   test('0-0', () => {
