+++ tsconfig.json
@@ -1,9 +1,8 @@
 {
     "compilerOptions": {
       "esModuleInterop": true,
-      "module": "nodenext",
       "noEmit": true,
       "allowImportingTsExtensions": true,
-      "moduleResolution": "nodenext",
+      "moduleResolution": "node",
     }
 }
